JS and layout
The javascript code is changed. i made a code so the background color can change. i also made the layout and designed it. the only thing which does nog work right now is the background image and the media queries. i also have to work on showing the seconds by the time on my webpage.